requirements.txt
awscli
requests
apache-airflow
dlt
sqlalchemy>=1.4
pyodbc>= 4.0.0
snowflake-connector-python==3.7